+++ cf/regrid/regrid.py
@@ -14,21 +14,11 @@ from ..functions import DeprecationError, regrid_logging
 from ..units import Units
 from .regridoperator import RegridOperator
 
-# ESMF renamed its Python module to `esmpy` at ESMF version 8.4.0. Allow
-# either for now for backwards compatibility.
-esmpy_imported = False
+esmpy_imported = True
 try:
     import esmpy
-
-    esmpy_imported = True
 except ImportError:
-    try:
-        # Take the new name to use in preference to the old one.
-        import ESMF as esmpy
-
-        esmpy_imported = True
-    except ImportError:
-        pass
+    esmpy_imported = False
 
 logger = logging.getLogger(__name__)