+++ cf/test/test_weights.py
@@ -217,20 +217,15 @@ class WeightsTest(unittest.TestCase):
         self.assertTrue((w.array == correct_weights).all())
         self.assertEqual(w.Units, cf.Units("m2"))
 
-        # Check that the global sum of cell areas is equal between a
-        # global HEALPix grid (with a single refinment level) and its
-        # UGRID version. This test assumes that the HEALPix areas are
-        # correct.
-        h = cf.example_field(12)
-        u = h.healpix_to_ugrid()
-        h_weights = h.weights(measure=True, components=True)[(1,)]
-        u_weights = u.weights(measure=True, components=True,great_circle=True)[(1,)]
-        global_area = float(4 * np.pi * (h.radius()**2))
-        h_area = h_weights.sum().array
-        u_area = u_weights.sum().array
-        self.assertTrue(np.allclose(h_area,  global_area))
-        self.assertTrue(np.allclose(h_area, u_area))
-        
+        # For a UGRID derived from a global HEALPix grid, check that
+        # the global sum of cell areas is correct
+        u = cf.example_field(12).healpix_to_ugrid()
+        u_weights = u.weights(
+            measure=True, components=True, great_circle=True
+        )[(1,)]
+        global_area = 4 * np.pi * (u.radius() ** 2)
+        self.assertTrue(np.allclose(u_weights.sum(), global_area))
+
     def test_weights_line_length_geometry(self):
         # Spherical line geometry
         gls = gps.copy()
@@ -361,10 +356,11 @@ class WeightsTest(unittest.TestCase):
         self.assertTrue(np.allclose(w[16:], 1 / (4**1)))
 
         w = f.weights(measure=True, components=True)[(1,)].array
-        radius = f.radius()
-        x = 4 * np.pi * (radius**2) / 12
+        x = 4 * np.pi * (f.radius() ** 2) / 12
         self.assertTrue(np.allclose(w[:16], x / (4**2)))
         self.assertTrue(np.allclose(w[16:], x / (4**1)))
+        # Total global area
+        self.assertTrue(np.allclose(w.sum(), x * 12))
 
 
 if __name__ == "__main__":