+++ Changelog.rst
@@ -11,6 +11,8 @@ Version NEXTVERSION
   (https://github.com/NCAS-CMS/cf-python/issues/???)
 * New method: `cf.Data.coarsen`
   (https://github.com/NCAS-CMS/cf-python/issues/???)
+* New function: `cf.locate`
+  (https://github.com/NCAS-CMS/cf-python/issues/???)
 * New optional dependency: ``healpix>=2025.1``
 * Changed dependency: ``cfdm>=1.13.0.0, <1.13.1.0``